Rise quirk_t.type to 16 bit. This fixes 9496180ea0e7e549167c12b8f9fe85a26b91cce2

+++ system/hwquirks.h
@@ -38,7 +38,7 @@ typedef enum {
 
 typedef struct {
     quirk_id_t   id;
-    uint8_t      type;
+    uint16_t     type;
     uint16_t     root_vid;
     uint16_t     root_did;
     void (*process)(void);
